feat: moved algos to model & label collision fix is now more precise
