+++ src/main/kotlin/viewmodel/EdgeViewModel.kt
@@ -8,19 +8,23 @@ import model.DirectedGraph
 import model.Edge
 import model.UndirectedGraph.UndirectedEdge
 
-class EdgeViewModel<V, K, W: Comparable<W>>(
+private const val COLLISION_FIX = 5000
+
+class EdgeViewModel<V, K, W : Comparable<W>>(
 	val firstVertexViewModel: VertexViewModel<V>,
 	val secondVertexViewModel: VertexViewModel<V>,
 	model: Edge<V, K, W>,
 	color: Color,
 	var width: Dp,
 	private val weightLabelVisieState: State<Boolean>
-): Colorable<Edge<V, K, W>> {
+) : Colorable<Edge<V, K, W>> {
 
 	private val _model = mutableStateOf(model)
 	override var model: Edge<V, K, W>
 		get() = _model.value
-		set(value) { _model.value = value}
+		set(value) {
+			_model.value = value
+		}
 
 	private val _color = mutableStateOf(color)
 	override var color: Color
@@ -34,4 +38,12 @@ class EdgeViewModel<V, K, W: Comparable<W>>(
 
 	val weightLabelVisible
 		get() = weightLabelVisieState.value
+
+	override fun hashCode(): Int {
+		return ((firstVertexViewModel.hashCode() - secondVertexViewModel.hashCode())).rem(COLLISION_FIX)
+	}
+
+	override fun equals(other: Any?): Boolean {
+		return super.equals(other)
+	}
 }
